requirements.txt
flask==2.3.3
tensorflow==2.17.1
Pillow==9.3.0
numpy>=1.24.3
google-cloud-storage==2.10.0
python-dotenv==1.0.0
